Add uninstall, include checks, more info.  Add uninstall only param as well

+++ install.sh
@@ -41,12 +41,15 @@ function check_download {
 }
 
 function remove_links {
+    echo "[UNINSTALL] Remove old links..."
     for path in "${KLIPPER_PATH}"/klippy/extras/ "${CONFIG_PATH}"/config/; do
-        echo "[UNINSTALL] Old links in ${path}";
-        for file in $(find "${KLIPPER_PATH}"/klippy/extras -type l -exec readlink -f {} \; | grep -F "${HOME}"/klipper-toolchanger); do
-            rm ${file};
+        for file in $(find "${path}" -type l); do
+            if readlink -f "${file}" | grep -q "${INSTALL_PATH}"; then
+              rm ${file};
+            fi
         done
     done
+    echo
 }
 
 function link_extension {
@@ -54,6 +57,7 @@ function link_extension {
     for file in "${INSTALL_PATH}"/klipper/extras/*.py; do
         ln -sfn ${file} "${KLIPPER_PATH}"/klippy/extras/;
     done
+    echo
 }
 
 function link_macros {
@@ -61,6 +65,7 @@ function link_macros {
     for file in "${INSTALL_PATH}"/macros/*.cfg; do
         ln -sfn ${file} "${CONFIG_PATH}"/config/;
     done
+    echo
 }
 
 function copy_examples {
@@ -68,47 +73,55 @@ function copy_examples {
     for file in "${INSTALL_PATH}"/examples/*.cfg; do
         cp -n ${file} "${CONFIG_PATH}"/config/;
     done
+    echo
 }
 
 function check_includes {
     echo "[CHECK-INSTALL] Checking for missing includes..."
     for file in "${INSTALL_PATH}"/macros/*.cfg; do
         filename=$(basename ${file});
-        if !grep -q "include ${filename}" "${CONFIG_PATH}"/config/printer.cfg; then
+        if ! grep -q "include ${filename}" "${CONFIG_PATH}"/config/printer.cfg; then
             echo " - Missing [include ${filename}] in printer.cfg"
         fi
     done
+    echo
 }
 
 function restart_klipper {
     echo "[POST-INSTALL] Restarting Klipper..."
     sudo systemctl restart klipper
+    echo
 }
 
 printf "\n======================================\n"
 echo "- Klipper toolchanger install script -"
-echo "-                                    -"
-echo "- If you are upgrading maybe sure to -"
-echo "- you check for changes in the user  -"
-echo "- config files                       -"
 printf "======================================\n\n"
 
-install=1;
-if [ $1 == 'uninstall' ]; then
-    install=0;
+doinstall=1;
+if [ $# -gt 0 ]; then
+    if [ "$1" == "uninstall" ]; then
+        doinstall=0;
+    fi
 fi
 
 # Run steps
-if [ $install -eq 1]; then
+if [ $doinstall -gt 0 ]; then
     preflight_checks
     check_download
 fi
 
 remove_links
 
-if [ $install -eq 1]; then
+if [ $doinstall -gt 0 ]; then
     link_extension
     link_macros
     copy_examples
+    check_includes
     restart_klipper
+
+    printf "======================================\n"
+    echo "- If you are upgrading maybe sure to -"
+    echo "- you check for changes in the user  -"
+    echo "- config files                       -"
+    printf "======================================\n\n"
 fi
